bruteforce/bruteforceWednesday/requirements.txt
# SSH Brute Force Detection & APT Correlation System
# Python Dependencies

# Core Data Processing
pandas>=1.3.0

# Machine Learning
scikit-learn>=1.0.0
joblib>=1.1.0

# Kafka Integration
kafka-python>=2.0.0

# Date/Time Processing
python-dateutil>=2.8.0

# Built-in libraries (included for completeness, no installation needed):
# - re (regex operations)
# - os (operating system interface)
# - datetime (date and time handling)
# - collections (specialized container datatypes)
